% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean data by removing NA Values}
\usage{
clean_data(data, remove_na = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{remove_na}{Logical. If TRUE, remove rows with NAs.}
}
\value{
A cleaned data frame
}
\description{
Clean data by removing NA Values
}
