% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predictions.R
\name{plot_predictions}
\alias{plot_predictions}
\title{Plot Actual vs Predicted Values}
\usage{
plot_predictions(model, data, target)
}
\arguments{
\item{model}{A trained model object.}

\item{data}{A data frame containing the test data.}

\item{target}{The target variable (string).}
}
\value{
A ggplot object showing the plot.
}
\description{
Plot Actual vs Predicted Values
}
