% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Split data into training and test sets}
\usage{
split_data(data, target, prop = 0.8)
}
\arguments{
\item{data}{A data frame.}

\item{target}{The target variable as a string.}

\item{prop}{The proportion of data to use for training.}
}
\value{
A list with two data frames: $train and $test.
}
\description{
Split data into training and test sets
}
