% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_model}
\alias{train_model}
\title{Train a Predictive Model}
\usage{
train_model(data, target, method = "lm")
}
\arguments{
\item{data}{A data frame with training data.}

\item{target}{The target variable string.}

\item{method}{The model method ("lm", "rf", "svm")}
}
\value{
A trained model object.
}
\description{
Train a Predictive Model
}
